Use SplitFileReferences in InMemoryFileReferenceStoreTest
