+++ java/core/src/test/java/sleeper/core/statestore/inmemory/InMemoryFileReferenceStoreTest.java
@@ -27,6 +27,7 @@ import sleeper.core.statestore.AllFileReferences;
 import sleeper.core.statestore.FileReference;
 import sleeper.core.statestore.FileReferenceFactory;
 import sleeper.core.statestore.SplitFileReference;
+import sleeper.core.statestore.SplitFileReferences;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.StateStoreException;
 
@@ -136,7 +137,7 @@ public class InMemoryFileReferenceStoreTest {
 
     @Nested
     @DisplayName("Split file references across multiple partitions")
-    class SplitFileReferences {
+    class SplitFiles {
         @Test
         void shouldSplitOneFileInRootPartition() throws Exception {
             // Given
@@ -145,8 +146,7 @@ public class InMemoryFileReferenceStoreTest {
             store.addFile(file);
 
             // When
-            store.splitFileReferences(List.of(
-                    splitFileToChildPartitions(file, "L", "R")));
+            SplitFileReferences.from(store).split();
 
             // Then
             assertThat(store.getActiveFiles())
@@ -164,9 +164,7 @@ public class InMemoryFileReferenceStoreTest {
             store.addFiles(List.of(file1, file2));
 
             // When
-            store.splitFileReferences(List.of(
-                    splitFileToChildPartitions(file1, "L", "R"),
-                    splitFileToChildPartitions(file2, "L", "R")));
+            SplitFileReferences.from(store).split();
 
             // Then
             assertThat(store.getActiveFiles())
@@ -188,9 +186,7 @@ public class InMemoryFileReferenceStoreTest {
             store.addFiles(List.of(file1, file2));
 
             // When
-            store.splitFileReferences(List.of(
-                    splitFileToChildPartitions(file1, "LL", "LR"),
-                    splitFileToChildPartitions(file2, "RL", "RR")));
+            SplitFileReferences.from(store).split();
 
             // Then
             assertThat(store.getActiveFiles())
@@ -199,6 +195,52 @@ public class InMemoryFileReferenceStoreTest {
                             splitFile(file1, "LR"),
                             splitFile(file2, "RL"),
                             splitFile(file2, "RR"));
+        }
+
+        @Test
+        void shouldOnlyPerformOneLevelOfSplits() throws Exception {
+            // Given
+            splitPartition("root", "L", "R", 5L);
+            splitPartition("L", "LL", "LR", 2L);
+            splitPartition("R", "RL", "RR", 7L);
+            FileReference file = factory.rootFile("file.parquet", 100L);
+            store.addFile(file);
+
+            // When
+            SplitFileReferences.from(store).split();
+
+            // Then
+            assertThat(store.getActiveFiles())
+                    .containsExactlyInAnyOrder(
+                            splitFile(file, "L"),
+                            splitFile(file, "R"));
+        }
+
+        @Test
+        void shouldNotSplitOneFileInLeafPartition() throws Exception {
+            // Given
+            splitPartition("root", "L", "R", 5L);
+            FileReference file = factory.partitionFile("L", "already-split.parquet", 100L);
+            store.addFile(file);
+
+            // When
+            SplitFileReferences.from(store).split();
+
+            // Then
+            assertThat(store.getActiveFiles())
+                    .containsExactly(file);
+        }
+
+        @Test
+        void shouldDoNothingWhenNoFilesExist() throws StateStoreException {
+            // Given
+            splitPartition("root", "L", "R", 5);
+
+            // When
+            SplitFileReferences.from(store).split();
+
+            // Then
+            assertThat(store.getActiveFiles()).isEmpty();
         }
 
         @Test
@@ -227,9 +269,7 @@ public class InMemoryFileReferenceStoreTest {
             store.addFile(file);
 
             // When / Then
-            assertThatThrownBy(() ->
-                    store.splitFileReferences(List.of(
-                            splitFileToChildPartitions(file, "L", "R"))))
+            assertThatThrownBy(() -> SplitFileReferences.from(store).split())
                     .isInstanceOf(StateStoreException.class);
             assertThat(store.getActiveFiles()).containsExactly(
                     file,
@@ -809,8 +849,9 @@ public class InMemoryFileReferenceStoreTest {
         }
     }
 
-    private void splitPartition(String parentId, String leftId, String rightId, long splitPoint) {
-        partitions.splitToNewChildren(parentId, leftId, rightId, splitPoint);
+    private void splitPartition(String parentId, String leftId, String rightId, long splitPoint) throws StateStoreException {
+        partitions.splitToNewChildren(parentId, leftId, rightId, splitPoint)
+                .applySplit(store, parentId);
         factory = FileReferenceFactory.fromUpdatedAt(partitions.buildTree(), DEFAULT_UPDATE_TIME);
     }
